Commit 2
It seems that I have fixed the problem of identifying too few groups. I did this by integrating over the betas when a new group is created. This avoids the problem of having wild betas sampled from the prior when a group is empty, which has the implication that new groups are never formed.